Fixed jsonobject memory leaks: is not truly manageable by intrusive_ptr, switched back to shared_ptr.
Problem is that Json-C json_objects can be independently owned by Json-C, bypassing the JsonObject wrapper. But when the JsonObject wrapper uses the embedded Json-C object's refcount, it will not get disposed as long as the Json-C is still owned otherwise, and when that ownership is released directly (by Json-C internal), the wrapper will leak.