macaddress: now should work for network interface configurations with gaps in interface index series.
The old assumption (across the internet, including accepted Stackoverflow answers) was that network interface indices are always counting up from 1.
This is no longer true (probably never really was), because dynamic adding and removal of network interfaces and dependencies can lead to gaps.

Workaround is not perfect, if the interface to be used as IP address or MAC address base has an index>20 *and* there are gaps above index 20, that interface would not be found.