mainloop: thread signal handler treated all signals as indication of thread ending. This was only correct before httpThreadSignalDataReady was introduced.
When httpThreadSignalDataReady was used in stream-returning http request, it caused the self-retaining (selfRef) of the child thread wrapper object get released too early (at the first occurrence of httpThreadSignalDataReady, or any other non-terminating signal). As the child thread object usually is kept alive by other pointers, a problem only occurred later, when, in the process of ending the request, all those other pointers released the wrapper object (as part of callback execution) and caused its data to be freed *before* signalPipeHandler() actually completed.