i2c PCA9685: found that on-time shifting does cause flickering when off-time wraps around -> disabled shifting for now.
It is unclear why this happens, all data sent to the PCA9685 looks ok down to the SMBus transactions, but fact is that transitioning from a non-wrapped around off-time to a wrapped around one causes flickering, i.e. periods where the output is off for a visible amount of time before recovering. Would need low level analysis with a scope. For now, disabling shift fixes the problem at the cost of all PWM on edges being at the same point in time, thus demanding a bit more from the supply.