civetweb: made build again with NO_SSL_DL undefined
Means that all libssl/libcrypto functions need to be defined as macros referring to an entry in the global function tables populated with function pointers when SSL is actually needed. Host verification uses a few more fumctions, so these had to be added.

Note GENERAL_NAME_free is a special case, which is invoked via a macro that does not allow to substitute the name with a function table pointer expression.