mainloop: prevent invalid iterator access that could happen while processing a destruction chain.
Thanks Michal for the precise analysis of this tricky issue!

The problem was that the else block local runningTimer went out of scope and could trigger a destruction chain making timer changes *after* the timersChanged flag was checked for the next iteration of the loop.

By moving the check to the top of the loop it is ensured all destructions happen before the check.