LedChainComm: multiple areas can be mapped onto the same ledchain now.
On the LedChainComm level, use setChainDriver() to act only as a coordinate mapper and delegate actual output to hardware to another LedChainComm doing the actual output to hardware.

On the LedChainArrangement level, adding multiple chains with the same driver name makes 2nd and further chains act as secondary mapper on the first chain.

This allows to use a single chain covering more than only one rectangle of an arrangement.