rpi_ws281x: updates from upstream github.com/jgarff/rpi_ws281x (mostly: 64bit and RPi4 support)
Note: this removes my own change in mailbox.c that differentiated between kernel versions to use either 100 or 249 for devmaj of the mailbox file node, as long time ago 100 did not work on >=v4.1 kernels. Most probably, this problem does not exist any more (otherwise, other people would have found it), so going back to the original devmaj 100 for now.