Logger: remove loglevel checks in most actual logging methods, use logEnabled checks (as LOG macros do) instead
Note: reason for the change is the new log level offset parameter in logEnabled(), which allows elevating/silencing log info per logging source. So the loglevel passed to output functions (except log()) no longer determines if the message should be shown or not, but only the level to show in the message prefix.