SocketComm: p44script support for receivin UDP messages as events
Note: current architecture limitations requires the udpsocket.message() assigned to a global var in order to be used in a on() handler, rather than writing "on(socket.message())". This is ugly and non-intuitive, but at least script done this way will not break once a better event architecture is in place in the future.