esp32_ws281x: finally, flicker free WS281x RMT data output!
Problem fixed now that still caused flicker in previous version:

After the initial load of both buffer halves, when the first THR-IRQ was too late, RMT output re-sent first half of the buffer (because no stopper present), causing ugly flicker.
Now we fill the second half of the buffer *after* starting RMT processing the first half, which includes placing a stopper in the first word of the first half. By the time this stopper is placed, the RMT is already past that first word, so the stopper does not disturb output of that half still in progress. To guarantee the stopper will be there in time in *all* cases, we need to block IRQs before starting RMT until after second half + buffer is written. Otherwise, a IRQ could steal too much time such that the stopper would not be in place early enough.