dcmotor: p44script: refactor motor state to be freezable (but no longer oneshot or keeporiginal)
actual motor status is now sampled at status object creation, and no longer when accessing jsonValue. The latter was done only to postpone status creation for efficiency reason to when it is actually used in a `on() { }` body, i.e. avoid creating the status twice. Now with freezing this is no longer a problem.