p44script: add `nowait` object type attribute, make await() check for it
Before, some objects (animators, threads) dynamically exposed a EventSource depending on internal activity in progress, to make await() exit immediately when applied to a non-active object.
But this prevented using these objects in `on() {}` as they would not register to trigger's EventSink when inactive.

The `nowait` flag is allowed to be dynamic, as it is checked only at await().