mainloop: cross thread call mechanisms
Main goal is to allow a subthread starting the crossThreadCallProcessor(), which then
accepts calls from the main thread. Short calls can be executed synchronously with the
main thread, but primary use case is running something that blocks for a longer time
on the subthread with the semantics of an async call with a callback parameter.

Implementation is such that calling back into the main threads main loop uses the
signal pipe mechanism already established for communicating threads ending.

Future implementation might add handling calls not only via crossThreadCallProcessor(),
but via a child thread mainloop. For that, we are currently missing a signal pipe
or similar mechanism to wake the child mainloop, like we have it from child to main already.